Add pages blueprint, content pages, localization wrapper, traffic page
Added a blueprint for all content pages (static stuff).
It is possible to add an english version of a template under
'templates/en/<name>.html', which gets called if 'lang' is 'en' in the
session settings and if the translated template exists.
If this is not the case the german version is returned.

As a first test the traffic FAQ is added too.

fixes #533