Insert zero traffic into where none was in the DB
As the traffic system did not log these days, there was only one row
returned by the query which asks for a user's week-old traffic log.
As a direct consecuence, only one row was displayed in the UI.
Now it is looping over each of the last week's days displaying a 0.0 MB
traffic in the UI when no entry has been returned.