Remove custom flat pages HTML renderer
The custom HTML renderer tried to ensure proper escaping of HTML
inside Markdown files. It didn't provide the same functionality
as the default pygements_markdown renderer, namely extensions, and
used a deprecated method to escape HTML. Escaping of HTML is not so
important as people with commit access to the commit repository
can already upload arbitrary files and must be therefore be trusted
anyway.