Setup proper package layout for the project
Python software should be placed in a proper top-level package
and shouldn't scatter its contents in several generic packages
in the top level namespace. We do have packages like utils
in the global namespace. This is bad. This has already been
done partially in a previous commit, but was reverted, probably
due to lack of understanding the issue.

This reverts commit 594ceb3f2db795711d409b9cc8e52a51ce10c161.