Configure a timeout for SQL connections
Fixes #590@10m

Although this fixes #590, one might want to handle exceptions different
dependent on if it is a connection error to Atlantis or Helios.
If only Helios is down, the sweb should still be able to display the
rest of the information and flash/log. But that is a lower priority
problem