Improve usability of /usertraffic info messages
Because /usertraffic displays the traffic data bound to the *ip* of the
device accessing the site, it might be confusing for a user who does not
know this. Thus, I added messages informing the user that e.g. he can't
see this page because he is not inside the wundtstraße right now,
*redirecting to /login*, even warning in the situation where the current
user *differs* from the user registered to this ip (although this is
pretty unreasonable, I admit).

I extracted a new method user_id_from_ip() from get_usertraffic()
because this information is now used in checking for the cases mentioned
above.