Complete name refactoring
In the previous commit, commit:5218c21b7a614110ff4ad8ecf12955340903f6b4,
just things in the code like refs and variable names have been refactored.
This completes this to the makefile, the readme and similiar.

PS: I need to write faster, just before I tried to push, @toothstone was
faster and fixed the paths in the makefile. Damnit!