Remove panel header of .md pages
Because the `title` property in the metadata of the .md file is mainly
used for the navigation bar, the actual headline will (or “should”)
appear in the body of the file – perhaps in a slightly longer form.

Therefore, there was no need for the title to appear on top of the
content. Also, the latter was displayed rather small which looked a bit
strange and inconsistent with even a 3rd layer headline.
It now is mentioned in the footer as “menu title”.