changed the system of config import and had to change application design
The Aim was to read each configuration for the utilis from app.config
That would mean in our old setup that we would have circular imports.
Now we fixed this at the moment with putting only the app in __init__.py
and loading and configuring the app with base.py.
This base.py loads at the end the views.py which loads all utilities.

The second problem was then that database_utils setup engines that needed a
configured app object.  So no database_utils depending moduls are now loaded
when we run init_app. That's not nice and has to be fixed in some way.