Move default config file location to /tmp to avoid access problems
Because currently, tha default logfile is of almost no importance, I moved it to
`/tmp` because without further doing, there will occur write access errors if
one is mounting `./` into the docker container, and this neds a quick fix.

The only thing that currently matters is a working stdout loghandler, as is
implemented already. Everything else is a question of deployment.