Global form redesign and global form rendering
The forms are now following:
http://getbootstrap.com/css/#forms-horizontal

The design is now enforced globally by creating a jinja macro in
`templates/macros/forms.html`. Custom data can be given in multiple
ways:

- Default values:  Can be assigned by `form.field.default = val`

- Placeholders:  Can be assigned using `form.field.description`

- Hints above the submit buttons:  Can be inserted using the caller.

Note there are two macros, `render` and `render_simple`.  The latter
dose not have a caller and can be invoked using `{{ render_simple() }}`.
A crucial difference is that `render_simple` has to get _keyword
arguments_ and does not work with positional ones.  This is due to a
design feature in jinja2 which does not enable passing of arguments in
the usual way of `def method(*args, **kwargs): call(*args, **kwargs);`.
Instead, `kwargs` is available as a global dict (not so `args`).

Labels have to be given in the form class definition as well.  Where
missing, the labels have been wrapped with`lazy_gettext()`.  Prior,
although the labels have been set for the fields in the class, they haev
been rendered redundantly with a completely different string in the
jinja template itself.

To sum up, weight has been reduced in the jinja templates and moved to
the form class definitions, as they are already providing structure.