Fix order of form fields
As the order of the fields is not defined statically in the templates
anymore, the order of definition in the field is important.

Note that `LoginForm.division` had to be rewritten such that it is not a
`LocalProxy` returning a `SelectField` anymore, but a `SelectField` in
the first place using a `LocalProxy` to define the choices.  The first
confuses the field ordering of the form, since it is technically of type
`LocalProxy` and not of type `BaseField` (or however the base class is
called).