Use extension patern for active divisions
The “active divisions” – which refers to the divisions which are not
deactivated due to `division.debug_only`, except if the application runs
in debug mode, of course – are now being put into
`current_app.extensions['divisions']`.

The initialization, which is implemented in `model.init_divisions` takes
care of the destinction regarding the debug mode.  After it is called,
which is done in `init_app` before calling `init_context`, the divisions
are accessible in the extension as a list.