Manually pass debug switch to app object
The initialization process for forking environments and direct calling
is handled differently.  In the latter case, if given, the `app.debug`
option is enabled manually. Although this is given in `app.run()` as
well, `init_app()` is called before that, and since the distinction of
which division to mark as active depends on `app.debug`, this is
neccessary.