Choose dormitories instead of divisions
Note that the change is purely cosmetic.  In order to suggest the
dormitory based on the User's IP address as noted in #69, the
information regarding the subnets has to be moved into the `Dormitory`
class.  Note further that this would imply a one-to-one-correspondance
between a subnet and a dormitory, which makes it impossible to
distinguish between i.e. the ZW41* dormitories as they share one subnet.

Fix #56