Fix ip resolution in `model.wu`
Instead of passing the MySQL `user_id`, the LDAP `uid` has been passed
two times.

This lead to displaying the wrong credit, since the `user_id` is
expected to be an integer.  Because a string was given, it was cast to a
`0`, which is a valid user in the database though.  This explins why
there was no error, but just wrong data.

Fix #62