Make `Your email address` field readonly
Unfortunately, this requires a small hack, because in WTForms, one
cannot directly pass rendering arguments (such as `readonly=True`).
Although, this feature has been added six(!) days ago into the source
code, so it will probably be accessible when WTForms 3.0 is released.

Worth noting: obviously, the content of the field is not accessed, but
rather the default mail address determined earlier.

Refs #70