Fix implementation of unsupported dormitories
The unsupported dormitories are now introduced as actual dormitories
with datasources being instances of the new class `PrematureDataSource`,
whihc provides all the necessary information.  The core advantage is
that a mutual property of both `PrematureDataSource` and `Datasource` is
`support_mail`, which makes the integration of unsupported dormitories
into the anonymous contact form trivial.

Fix #123