Prepare `Traffic` for usage
The current implementation of `traffic.tuext` would fail, since this is
not only a table with a certain tablename, but a connection to a
different database (at least when using ORm abstraction).  Thus,
`Traffic` has to be bound to a different engine, which is done in
`init_db` when creating `Session`.