Split up + document `add_article`
`add_article` was not really in the right place.  What it did was, given
a filename, to add this as a localized page to an article.  If the
article of this name did not exist yet, a new `Article` was constructed
and given this page as the default.

The addition of the localized pages to an article is not the job of
`Category`, which is why this has been partly implemented in `Article`.

Just take a look at those beautiful docstrings :-)