Add json testing && improve api mocking
Since we don't actually care about the implementation (“was `.json`
called?”, “Was a `ValueError` catched?”, …), we don't need to mock the
`Response` object and all of its functions.

The better approach is to let the mocked `get`-function return an actual
`Response` object, and setting the binary content (`content` is a
property, thus `_content` is set).

Now, the test only has to verify whether the binary represantation of a
json-dump of a dict is being read into the same dict again.