Allow empty redirection mail address
In the gerok module, the `mail` field of the API's `userdata` JSON-dict
is supposed to be the redirection mail address.  Thus, if it is empty,
the `user.mail` property should contain the builtin mail address of the
gerok module.

The tests were extended so this usecase is in `assert_userdata_passed`.