Move build files to own dir, optimize `Dockerfile`
Now, requirements, docker-compose-`yml`s, and other stuff is in
`build`.  Important files have been symlinked to the top level.

The dockerfile has been optimized:  It now adds `./build` first and
installs the requirements *before* adding all the other files, so in
most cases, they don't have to be installed again, because the cache can
be used.  Previously, adding `/home/sipa/sipa` in advance caused the
cache to drop after every single code change, which cascaded to the
later installation of the requirements.