Modify `property` to support `raw_value` passing
Since the `money` decorator passes the original float as a `raw_value`,
this has to be supported.

When Implementing, this broke the tests, since they tested for equality
with the format string – however, `__eq__` of `ActiveProperty` compares
first with `raw_value`, if the compared object is no `ActiveProperty` –
which now is `3.5` instead of `+3.50 €`.

This simplifies the tests a lot, and is way more consistent with the
semantic meaning of `raw_value`.