Fix wu schema
The relationship-join that was done before should never have worked,
since it reached over separate databases.  It has been replaced by a
property `Nutzer.transactions` which directly executes a query,
correcting the value concerning where the `nutzer_id` stands with in the
transaction, and converting to €.

Plus according changes/fixes to the tests.