Fix config regarding prefixes
Currently, because of the coupled usage of `wsgi-file` and `mount`,
there have been two apps the whole time.  If `--mount <mount>` has been
given in the command line as well, attempting to override it, a third
uwsgi app joins the happy couple.

I don't know it caused the problem I wanted to fix (hyperventilating
timer calls), but that's bad anyway.

This commit modifies the `uwsgi.ini`, adds some documentation, and
modifies `development.yml` to operate correctly (and not set an obsolete
environment variable).

Fix #140 (finally)