Make datasource logic a `Backends` flask extension
Since the methods of `model.__init__` use the common state of the
available dormitories and datasources, it is simpler to introduce a new
class called `Backends` and refer to `self` instead of
`current_app.extensions` all the time.  This class is then registered as
the app's extension via `init_app`.

It provides an initialization method via `init_backends`.

The old methods have been deleted, the usages will be fixed in the next
few commits.