Automatically register datasources' dormitories
Since whenever you instanciate `Dormitory`, you must give a datasource,
it is useful to tell the datasource to “remember” this instanciation.
In this case, this means adding it to a dict.

This makes it easier to lookdatasources up or get a datasource by name,
and provide name consistency when doing the same with datasources.

Also, the exposure of the explicit `dormitories` lists in each model is
discontinued.