Register backends successively, check for sanity
Instead of hard-coding a list of datasources and dormitories,
just have a list of available (i.e. “implemented”) datasources and add
each entry successively.

Which datasources are actually used *must be activated manually* – else,
a warning is thrown and the dormitories aren't available.

Introduce functions for said successive registering, in which the
`dsrc.name` uniqueness check is done.  When registering a datasource,
it's dormitories are registered as well, following the same principle.