Extract contact mail composition
The responsibility of the view function `contact` only lies in sending
the mail and creating user feedback, not in composing the mail from its
pieces.

This logic has been extracted into a `send_contact_mail` function in
mail_utils.

Note that this should follow for other views in a similar manner.