Change model to rely on `user.finance_information` only
This introduces some changes to the way finance_information is
propagated in the model.

`BaseUser` now expects the class to implement the `finance_information`
property or return `None` if unsupported, which is meant to be a
`BaseFinanceInformation` instance.  Since every row must be a direct
attribute of the user (this restriction comes from the design with
`generate_rows`), the `BaseUser` provides `finance_balance` pointing to
`finance_information.balance` if supported (info being `True`ish).

The few tests for `BaseUser` which have been written (and added) have
been moved to `tests.model.test_user`.

This change does not include adapting all the backends except for
`sample`.