Use ProxyFix from werkzeug
The ReverseProxied middleware introduced in #31 to make reverse proxying
transparent for the WSGI app, is not complete. It uses uncommon headers
(X-Scheme, X-Script-Name) and does not handle X-Forwarded-For, although it
occurs in its docstring.

A better solutions is werkzeug.contrib.fixers.ProxyFix. It handles the
following headers:
* X-Forwarded-For
* X-Forwarded-Proto
* X-Forwarded-Host

The configuration of the upstream reverse proxies has to be adjusted, when
deploying the change. The script name is kept as is, because Flask can handle
different URL prefixes.