Fix membership evaluation in HSS
The `active` flag previously used indicates if a membership is active or
passive. It was used wrongly to check if a user logged in to SIPA is
still a member of AGDSN. However, in Abe it is common practice to delete
a person's account if they leave AGDSN. Therefore, it is a maybe not so
nice, but sound solution to just return true here.