Properly separate the Backends flask extension from its usage
`backends` is now its own package, and the `model` only contains the
code utilizing it.

In the process, some dependencies could be cleared up.  Registering
datasources now has to be done explicitly, and initializing the
sqlalchemy binds before init has to be registered via a hook, and not
hard-coded into the backends logic.