Put `<script>` tags into the `head`, using `defer`
Putting script tags at the end of the HTML element was done to avoid
blocking further HTML parsing and DOM loading – whenever a script tag
is encountered, a browser has to parse and interpret the JS inside of
it.  A disadvantage of this strategy is that the scripts will be fetched
only after parsing the HTML

Using `defer` while putting the scripts in the `head` solves both of
these problems.